+++ profiles/bicycle.lua
@@ -401,7 +401,7 @@ function way_function (way, result)
   limit( result, maxspeed, maxspeed_forward, maxspeed_backward )
 end
 
-function turn_function (angle)
+function turn_function (angle, approach_road_speed, exit_road_speed)
   -- compute turn penalty as angle^2, with a left/right bias
   -- multiplying by 10 converts to deci-seconds see issue #1318
   k = 10*turn_penalty/(90.0*90.0)